Move SensorSet into sensor subpackage
SensorSet is now independent of any concept of connections, and
instead has a notification mechanism that the device server uses
to dissociate sensors from connections when they're removed.
