+++ aiokatcp/server.py
@@ -1,4 +1,4 @@
-# Copyright 2017 National Research Foundation (Square Kilometre Array)
+# Copyright 2017, 2019 National Research Foundation (Square Kilometre Array)
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions are met:
@@ -30,16 +30,13 @@ import asyncio
 import functools
 import logging
 import traceback
-import enum
 import io
 import re
 import time
-from typing import (
-    Callable, Awaitable, Sequence, Iterable, Mapping,
-    Iterator, Optional, List, Tuple, Any, TypeVar,
-    Union, KeysView, ValuesView, ItemsView, cast, overload)
+from typing import (Set, Callable, Awaitable, Sequence, Iterable,
+                    Optional, List, Tuple, Any, TypeVar, cast)
 # Only used in type comments, so flake8 complains
-from typing import Dict, Set    # noqa: F401
+from typing import Dict    # noqa: F401
 
 import aiokatcp
 from . import core, connection, sensor
@@ -192,125 +189,6 @@ class DeviceServerMeta(type):
         return result
 
 
-class SensorSet(Mapping[str, sensor.Sensor]):
-    """A dict-like and set-like collection of sensors.
-
-    It stores a corresponding list of connections, and removing a sensor from
-    the set clears any samplers on the corresponding connection.
-    """
-
-    class _Sentinel(enum.Enum):
-        """Internal enum used to signal that no default is provided to pop"""
-        NO_DEFAULT = 0
-
-    def __init__(self, connections: Set[ClientConnection]) -> None:
-        self._connections = connections
-        self._sensors = {}       # type: Dict[str, sensor.Sensor]
-
-    def _removed(self, s: sensor.Sensor):
-        """Clear a sensor's samplers from all connections."""
-        for conn in self._connections:
-            conn.set_sampler(s, None)
-
-    def add(self, elem: sensor.Sensor):
-        if elem.name in self._sensors:
-            if self._sensors[elem.name] is not elem:
-                del self[elem.name]
-            else:
-                return
-        self._sensors[elem.name] = elem
-
-    def remove(self, elem: sensor.Sensor) -> None:
-        if elem not in self:
-            raise KeyError(elem.name)
-        del self[elem.name]
-
-    def discard(self, elem: sensor.Sensor) -> None:
-        try:
-            self.remove(elem)
-        except KeyError:
-            pass
-
-    def clear(self) -> None:
-        while self._sensors:
-            self.popitem()
-
-    def popitem(self) -> Tuple[str, sensor.Sensor]:
-        name, value = self._sensors.popitem()
-        self._removed(value)
-        return name, value
-
-    def pop(self, key: str,
-            default: Union[sensor.Sensor, None, _Sentinel] = _Sentinel.NO_DEFAULT) \
-            -> Optional[sensor.Sensor]:
-        if key not in self._sensors:
-            if isinstance(default, self._Sentinel):
-                raise KeyError(key)
-            else:
-                return default
-        else:
-            s = self._sensors.pop(key)
-            self._removed(s)
-            return s
-
-    def __delitem__(self, key: str) -> None:
-        s = self._sensors.pop(key)
-        self._removed(s)
-
-    def __getitem__(self, name: str) -> sensor.Sensor:
-        return self._sensors[name]
-
-    @overload
-    def get(self, name: str) -> Optional[sensor.Sensor]: ...
-
-    @overload     # noqa: F811
-    def get(self, name: str, default: Union[sensor.Sensor, _T]) -> Union[sensor.Sensor, _T]: ...
-
-    def get(self, name: str, default: object = None) -> object:    # noqa: F811
-        return self._sensors.get(name, default)
-
-    def __contains__(self, s: object) -> bool:
-        if isinstance(s, sensor.Sensor):
-            return s.name in self._sensors and self._sensors[s.name] is s
-        else:
-            return s in self._sensors
-
-    def __len__(self) -> int:
-        return len(self._sensors)
-
-    def __bool__(self) -> bool:
-        return bool(self._sensors)
-
-    def __iter__(self) -> Iterator[str]:
-        return iter(self._sensors)
-
-    def keys(self) -> KeysView[str]:
-        return self._sensors.keys()
-
-    def values(self) -> ValuesView[sensor.Sensor]:
-        return self._sensors.values()
-
-    def items(self) -> ItemsView[str, sensor.Sensor]:
-        return self._sensors.items()
-
-    def copy(self) -> Dict[str, sensor.Sensor]:
-        return self._sensors.copy()
-
-    __hash__ = None     # type: ignore     # mypy can't handle this
-
-    add.__doc__ = set.add.__doc__
-    remove.__doc__ = set.remove.__doc__
-    discard.__doc__ = set.discard.__doc__
-    clear.__doc__ = dict.clear.__doc__
-    popitem.__doc__ = dict.popitem.__doc__
-    pop.__doc__ = dict.pop.__doc__
-    get.__doc__ = dict.get.__doc__
-    keys.__doc__ = dict.keys.__doc__
-    values.__doc__ = dict.values.__doc__
-    items.__doc__ = dict.items.__doc__
-    copy.__doc__ = dict.copy.__doc__
-
-
 class DeviceServer(metaclass=DeviceServerMeta):
     """Server that handles katcp.
 
@@ -395,7 +273,14 @@ class DeviceServer(metaclass=DeviceServerMeta):
         self._host = host
         self._port = port
         self._stopping = False
-        self.sensors = SensorSet(self._connections)
+        self.sensors = sensor.SensorSet()
+        self.sensors.add_remove_callback(
+            functools.partial(self._remove_sensor_callback, self._connections))
+
+    @staticmethod
+    def _remove_sensor_callback(connections: Set[ClientConnection], sensor: sensor.Sensor):
+        for conn in connections:
+            conn.set_sampler(sensor, None)
 
     async def start(self) -> None:
         """Start the server running on the event loop.
