+++ aiokatcp/test/test_sensor.py
@@ -34,7 +34,7 @@ import unittest
 
 import asynctest
 
-from aiokatcp.sensor import Sensor, SensorSampler
+from aiokatcp.sensor import Sensor, SensorSampler, SensorSet
 
 
 class TestSensor(unittest.TestCase):
@@ -64,3 +64,158 @@ class TestSensorSampling(asynctest.TestCase):
             # Run gc twice because PyPy sometimes needs this.
             gc.collect()
             gc.collect()
+
+
+class TestSensorSet(unittest.TestCase):
+    def setUp(self):
+        self.ss = SensorSet()
+        self.callback = unittest.mock.MagicMock()
+        self.ss.add_remove_callback(self.callback)
+        self.sensors = [Sensor(int, 'name{}'.format(i)) for i in range(5)]
+        # A different set of sensors with the same names
+        self.alt_sensors = [Sensor(float, 'name{}'.format(i)) for i in range(5)]
+        self.ss.add(self.sensors[0])
+
+    def _assert_sensors(self, ss, sensors):
+        """Assert that `ss` has the same sensors as `sensors`"""
+        ordered = sorted(ss.values(), key=lambda x: x.name)
+        self.assertEqual(ordered, sensors)
+
+    def test_construct(self):
+        """Test that setUp put things into the right state."""
+        self._assert_sensors(self.ss, [self.sensors[0]])
+
+    def test_add(self):
+        # Add a new one
+        self.ss.add(self.sensors[1])
+        self._assert_sensors(self.ss, [self.sensors[0], self.sensors[1]])
+        # Add the same one
+        self.ss.add(self.sensors[0])
+        self._assert_sensors(self.ss, [self.sensors[0], self.sensors[1]])
+        # Replace one
+        self.callback.assert_not_called()
+        self.ss.add(self.alt_sensors[1])
+        self._assert_sensors(self.ss, [self.sensors[0], self.alt_sensors[1]])
+        self.callback.assert_called_once_with(self.sensors[1])
+
+    def test_remove(self):
+        # Try to remove non-existent name
+        with self.assertRaises(KeyError):
+            self.ss.remove(self.sensors[4])
+        # Try to remove one with the same name as an existing one
+        with self.assertRaises(KeyError):
+            self.ss.remove(self.alt_sensors[0])
+        self._assert_sensors(self.ss, [self.sensors[0]])
+        # Remove one
+        self.callback.assert_not_called()
+        self.ss.remove(self.sensors[0])
+        self._assert_sensors(self.ss, [])
+        self.callback.assert_called_once_with(self.sensors[0])
+
+    def test_discard(self):
+        # Try to remove non-existent name
+        self.ss.discard(self.sensors[4])
+        # Try to remove one with the same name as an existing one
+        self.ss.discard(self.alt_sensors[0])
+        self._assert_sensors(self.ss, [self.sensors[0]])
+        # Remove one
+        self.callback.assert_not_called()
+        self.ss.discard(self.sensors[0])
+        self._assert_sensors(self.ss, [])
+        self.callback.assert_called_once_with(self.sensors[0])
+
+    def test_clear(self):
+        self.ss.add(self.sensors[1])
+        self.ss.clear()
+        self._assert_sensors(self.ss, [])
+        self.callback.assert_any_call(self.sensors[0])
+        self.callback.assert_any_call(self.sensors[1])
+
+    def test_popitem(self):
+        self.ss.add(self.sensors[1])
+        items = []
+        try:
+            for _ in range(100):   # To prevent infinite loop if it's broken
+                items.append(self.ss.popitem())
+        except KeyError:
+            pass
+        items.sort(key=lambda x: x[0])
+        self.assertEqual(items, [('name0', self.sensors[0]), ('name1', self.sensors[1])])
+        self.callback.assert_any_call(self.sensors[0])
+        self.callback.assert_any_call(self.sensors[1])
+
+    def test_pop_absent(self):
+        # Non-existent name
+        with self.assertRaises(KeyError):
+            self.ss.pop('name4')
+        # Non-existent with defaults
+        self.assertIsNone(self.ss.pop('name4', None))
+        self.assertEqual(self.ss.pop('name4', 'foo'), 'foo')
+        # Remove one
+        self.callback.assert_not_called()
+        self.assertIs(self.ss.pop('name0'), self.sensors[0])
+        self._assert_sensors(self.ss, [])
+        self.callback.assert_called_once_with(self.sensors[0])
+
+    def test_delitem(self):
+        # Try to remove non-existent name
+        with self.assertRaises(KeyError):
+            del self.ss['name4']
+        self._assert_sensors(self.ss, [self.sensors[0]])
+        # Remove one
+        self.callback.assert_not_called()
+        del self.ss['name0']
+        self._assert_sensors(self.ss, [])
+        self.callback.assert_called_once_with(self.sensors[0])
+
+    def test_getitem(self):
+        # Non-existing name
+        with self.assertRaises(KeyError):
+            self.ss['name4']
+        # Existing name
+        self.assertIs(self.ss['name0'], self.sensors[0])
+
+    def test_get(self):
+        # Non-existing name
+        self.assertIsNone(self.ss.get('name4'))
+        self.assertIsNone(self.ss.get('name4', None))
+        self.assertEqual(self.ss.get('name4', 'foo'), 'foo')
+        # Existing name
+        self.assertIs(self.ss.get('name0'), self.sensors[0])
+
+    def test_len(self):
+        self.assertEqual(len(self.ss), 1)
+        self.ss.add(self.sensors[1])
+        self.assertEqual(len(self.ss), 2)
+
+    def test_contains(self):
+        self.assertIn(self.sensors[0], self.ss)
+        self.assertNotIn(self.alt_sensors[0], self.ss)
+        self.assertNotIn(self.sensors[1], self.ss)
+
+    def test_bool(self):
+        self.assertTrue(self.ss)
+        self.ss.clear()
+        self.assertFalse(self.ss)
+
+    def test_keys(self):
+        self.ss.add(self.sensors[1])
+        self.assertEqual(sorted(self.ss.keys()), ['name0', 'name1'])
+
+    def test_values(self):
+        self.ss.add(self.sensors[1])
+        self.assertEqual(sorted(self.ss.values(), key=lambda x: x.name),
+                         [self.sensors[0], self.sensors[1]])
+
+    def test_items(self):
+        self.ss.add(self.sensors[1])
+        self.assertEqual(sorted(self.ss.items()), [
+            ('name0', self.sensors[0]),
+            ('name1', self.sensors[1])
+        ])
+
+    def test_iter(self):
+        self.assertEqual(sorted(iter(self.ss)), ['name0'])
+
+    def test_copy(self):
+        self.assertEqual(self.ss.copy(), {'name0': self.sensors[0]})
